pyproject.toml
[project]
name = "nyc-yellow-taxi-mlops"
version = "0.1.0"
description = "NYC Yellow Taxi MLOps Pipeline with Airflow and Spark"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "apache-airflow>=2.8.2",
    "pyspark==3.4.1",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "boto3>=1.34.0",
    "s3fs>=2024.2.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyarrow>=14.0.0",
    "PyYAML>=6.0.1",
    "ruff>=0.15.0",
    "scikit-learn>=1.6.1",
    "optuna>=3.6.1",
    "lightgbm>=4.5.0",
    "joblib>=1.4.2",
    "mlflow>=2.17.0",
    "evidently>=0.7.20",
]

[project.optional-dependencies]
inference = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "prometheus-client>=0.20.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
